Reworked the project structure for better performance and thread safety

+++ src/pyodb/pyodb.py
@@ -1,15 +1,13 @@
 """Main module handling containing the main capabilities of the library.
 """
-import logging
 from pathlib import Path
 from time import time
 from typing import Any, Callable
 
-from src.pyodb._util import create_logger
-from src.pyodb.error import BadTypeError, CacheError, PyODBError
-from src.pyodb.schema.base._sql_builders import Delete, Select
-from src.pyodb.schema.shard_schema import ShardSchema
-from src.pyodb.schema.unified_schema import UnifiedSchema
+from pyodb.error import BadTypeError, CacheError
+from pyodb.schema.base._sql_builders import Delete, Select
+from pyodb.schema.shard_schema import ShardSchema
+from pyodb.schema.unified_schema import UnifiedSchema
 
 
 class PyODB:
@@ -51,25 +49,18 @@ class PyODB:
             Defaults to False.
         sharding (bool, optional): Whether to use sharding.
             Defaults to False.
-        write_log (bool, optional): Whether to enable logging.
-            Defaults to True.
-        log_to_console (bool, optional): Whether to output logs in the console as well.
-            Defaults to False.
         load_existing (bool, optional): Whether to load an existing schema or ignore it.
             Defaults to True.
     """
-    _logger: logging.Logger | None
     _schema: ShardSchema | UnifiedSchema
 
 
-    def __init__( # noqa: PLR0913
+    def __init__(
             self,
             max_depth: int = 2,
             pyodb_folder: str | Path = ".pyodb",
             persistent: bool = False,
             sharding: bool = False,
-            write_log: bool = True,
-            log_to_console: bool = False,
             load_existing: bool = True
         ) -> None:
         if not isinstance(pyodb_folder, Path):
@@ -84,40 +75,6 @@ class PyODB:
         if load_existing:
             self._schema.load_existing()
 
-        self.modify_logging(
-            do_logging=write_log,
-            log_folder=pyodb_folder,
-            console_output=log_to_console
-        )
-
-
-    def modify_logging(
-        self,
-        do_logging: bool = True,
-        log_level: int = logging.WARN,
-        log_folder: str | Path = ".pyodb",
-        console_output: bool = False
-    ):
-        """Modifies the libraries logging. Log files will have a maximum of 2MiB, with 2 rotating
-        files. So the logs will take up 6MiB at max.
-
-        Args:
-            log_level (int, optional): Defaults to WARN.
-            log_folder (str, optional): Defaults to "./logs". Folder must exist!
-            console_output (bool, optional): Whether to output logs in the console as well or only
-                to the file(s). Defaults to False.
-        """
-        if isinstance(log_folder, str):
-            log_folder = Path(log_folder)
-        log_folder.mkdir(mode=755, exist_ok=True)
-
-        if do_logging:
-            self._logger = create_logger(log_folder.as_posix(), log_level, console_output)
-        else:
-            self._logger = None
-        self._schema.logger = self._logger
-        PyODBError.set_logger(self._logger)
-
 
     @property
     def max_depth(self) -> int:
@@ -130,6 +87,18 @@ class PyODB:
         self._schema.max_depth = val
 
 
+    @property
+    def persistent(self) -> bool:
+        """Whether the database is persistent after closing.
+        In multiprocessing/threading this should always be set to true!"""
+        return self._schema.is_persistent
+
+
+    @persistent.setter
+    def persistent(self, val: bool):
+        self._schema.is_persistent = val
+
+
     def save(self, obj: object, expires: float | None = None):
         """Saves object to the database. Adds type in case it is not known.
 
@@ -139,12 +108,8 @@ class PyODB:
                 database. Defaults to None.
         """
         obj_type = type(obj)
-        if self._logger:
-            self._logger.debug(f"Saving object of type {obj_type}")
 
         if not self._schema.is_known_type(obj_type):
-            if self._logger:
-                self._logger.info(f"Adding new type '{obj_type}'")
             self._schema.add_type(obj_type)
         self._schema.insert(obj, expires)
 
@@ -241,18 +206,6 @@ class PyODB:
         return self._schema.is_known_type(type_)
 
 
-    @property
-    def persistent(self) -> bool:
-        """Whether the database is persistent after closing.
-        In multiprocessing/threading this should always be set to true!"""
-        return self._schema.is_persistent
-
-
-    @persistent.setter
-    def persistent(self, val: bool):
-        self._schema.is_persistent = val
-
-
 class PyODBCache:
     """Class that caches arbitrary data and returns cached data if available.
     Also updates the data if it is expired. Expiry times are set when adding a new cache.
@@ -264,10 +217,6 @@ class PyODBCache:
         persistent (bool, optional): Whether the database should be persistent after closing.
             Defaults to False.
         sharding (bool, optional): Whether to use sharding.
-            Defaults to False.
-        write_log (bool, optional): Whether to enable logging.
-            Defaults to True.
-        log_to_console (bool, optional): Whether to output logs in the console as well.
             Defaults to False.
     """
     class _CacheItem:
@@ -287,7 +236,7 @@ class PyODBCache:
             self.expires = 0
 
 
-        def get_data(self):
+        def get_data(self) -> list | None:
             if self.expires < time():
                 self.data = []
                 return None
@@ -302,10 +251,6 @@ class PyODBCache:
     _caches: dict[str, _CacheItem]
     _pyodb: PyODB
 
-    @property
-    def logger(self) -> logging.Logger | None:
-        return self._pyodb._logger
-
 
     @property
     def pyodb(self) -> PyODB:
@@ -319,29 +264,26 @@ class PyODBCache:
         return self._caches.copy()
 
 
-    def __init__( # noqa: PLR0913
+    def __init__(
             self,
             max_depth: int = 0,
             pyodb_folder: str | Path = ".pyodb",
             persistent: bool = False,
             sharding: bool = False,
-            write_log: bool = True,
-            log_to_console: bool = False
         ) -> None:
         self._pyodb = PyODB(
             max_depth=max_depth,
             pyodb_folder=pyodb_folder,
             persistent=persistent,
             sharding=sharding,
-            write_log=write_log,
-            log_to_console=log_to_console,
             load_existing=False
         )
+        self._pyodb._schema.save_table_defs = False
         self._caches = {}
 
 
     def cache_exists(self, cache_key: str) -> bool:
-        """Check the existance of a cache with key `cache_key`
+        """Check the existence of a cache with key `cache_key`
 
         Args:
             cache_key (str): Key of the cache to check for.
@@ -385,10 +327,14 @@ class PyODBCache:
                 raise CacheError(
                     f"Cache '{cache_key}' already exists! Use 'force=True' to suppress this error."
                 )
+            elif data_type != self._caches[cache_key].data_type:
+                    self.pyodb.remove_type(self.caches[cache_key].dataclass)
             else:
-                self.pyodb.remove_type(self.caches[cache_key].dataclass)
+                self._caches[cache_key].data_func = data_func
+                self._caches[cache_key].lifetime = lifetime
+                return
 
-        dataclass = type(f"PyODBCache_{data_type.__name__}", (), {
+        dataclass = type(f"PyODBCache_{cache_key}", (), {
                 "__init__": self._dataclass_constructor,
                 "__annotations__": {"data": data_type | None, "expires": float}
             }
@@ -396,8 +342,6 @@ class PyODBCache:
         globals().update({dataclass.__name__: dataclass})
         self.pyodb.add_type(dataclass)
         self._caches[cache_key] = self._CacheItem(data_func, data_type, lifetime, dataclass)
-        if self.logger:
-            self.logger.info(f"Added cache definition for '{cache_key}'")
 
 
     def get_data(self, cache_key: str, *args, **kwargs) -> list[Any]:
@@ -425,28 +369,25 @@ class PyODBCache:
         # Try to get data from in-memory cache
         data = cache.get_data()
         if data is not None:
+            print("Using cached data")
             return data
 
         # Try to get data from database
         db_res = self.pyodb.select(cache.dataclass).all()
         data = [dp.data for dp in db_res]
         if data:
+            print("Using db data")
             cache.set_data(data, db_res[0].expires)
             return data
 
         try:
+            print("Generating new data")
             data = cache.data_func(*args, **kwargs)
             expires = time() + cache.lifetime
             self.pyodb.save_multiple([cache.dataclass(dp, expires) for dp in data], expires)
             cache.set_data(data, expires)
-            if self.logger:
-                self.logger.debug(f"Refreshed cached {cache_key}")
             return data
         except Exception as err:
-            if self.logger:
-                self.logger.error(
-                    f"Failed to get/refresh datacache {cache_key}! Details: {err}"
-                )
             raise err
